Fix duplicate input class from object builder mixins
We Fix this by marking the old v0 mixins as `OldMixinTarget`
